* 2055_cleanup_master:
  Identation of some source files.
  UNDELFS: code cleanup
  avoid often memory alocation&release
  x86_64: Cleanup of code for master
  Cleanup: avoid warnings for mc-library on x86_64
  Avoid double check some panel values before call of select_item().
  Cleanup of code for master
  Ticket #2055: Cleanup before 4.7.1 and 4.7.0.3 releases

+++ lib/tty/color-internal.h
@@ -19,21 +19,24 @@
 extern gboolean use_colors;
 extern gboolean mc_tty_color_disable;
 
-typedef struct mc_color_pair_struct {
+typedef struct mc_color_pair_struct
+{
     const char *cfg;
     const char *cbg;
     int ifg;
     int ibg;
-    int pair_index;
+    size_t pair_index;
     gboolean is_temp;
 } tty_color_pair_t;
 
+/* *INDENT-OFF* */
 typedef enum {
     SPEC_A_REVERSE		= -100,
     SPEC_A_BOLD			= -101,
     SPEC_A_BOLD_REVERSE		= -102,
     SPEC_A_UNDERLINE		= -103
 } tty_special_color_t;
+/* *INDENT-ON* */
 
 const char *tty_color_get_valid_name (const char *);
 int tty_color_get_index_by_name (const char *);
